qrpa/pyproject.toml
[project]
name = "qrpa"
version = "0.1.0"
description = "fluent qa rpa project"
authors = [
    {name = "fluentqa", email = "fluentqa@163.com"},
]
dependencies = [
    "pandas>=2.2.2",
    "tabula-py>=2.9.3",
    "pytest>=8.3.2",
]
requires-python = ">=3.11"
readme = "README.md"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.3.2",
]
